test: add simple test case using Iris dataset
The test case preps the Iris data to produce a training and test dataset.
It then creates an MLP instance and trains it using a set hyperparameters that produces a decent result.
Finally, it performs a inference on the test data with a fwd pass, calculates accuracy of the predictions, and prints the results.